+++ sdk/options_test.go
@@ -1,24 +1,6 @@
-// MIT License
+// Copyright 2026 ConveyorQ
 //
-// Copyright (c) 2026 ConveyorQ
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
+// SPDX-License-Identifier: Apache-2.0
 
 package conveyor
 
@@ -36,6 +18,7 @@ func TestOptionsApply(t *testing.T) {
 		WithToken("secret"),
 		WithQueues(map[string]int{"critical": 6, "default": 3}),
 		WithConcurrency(20),
+		WithMinServerVersion("v1.2.0"),
 	} {
 		opt(settings)
 	}
@@ -43,6 +26,7 @@ func TestOptionsApply(t *testing.T) {
 	require.Equal(t, "secret", settings.token)
 	require.Equal(t, map[string]int{"critical": 6, "default": 3}, settings.queues)
 	require.Equal(t, 20, settings.concurrency)
+	require.Equal(t, "v1.2.0", settings.minServerVersion)
 }
 
 func TestEnqueueOptionsApply(t *testing.T) {